src/requirements.txt
fastapi==0.95.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
uvicorn==0.21.1
SQLAlchemy==1.4.44
psycopg2==2.9.6
